feat: allow selectively disabling duplicate route detection
Duplicate route detection is enabled by default, via a flag in the `RouteCollector` class.
If the flag is set, but no duplicate route detector is present in the instance, one is created on the first request to detect duplicate routes.
If disabled, duplicate route detection is not performed.

The `RouteCollector` factory now also checks for a configuration value, `Mezzio\Router\RouteCollector.detect_duplicates`; when present, the value is passed for the optional second constructor argument of the `RouteCollector`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>